requirements.txt
# Background Coding Agents for Manufacturing
# Python dependencies

# Core dependencies
pyyaml>=6.0.1           # Configuration files
asyncio>=3.4.3          # Async operations

# Logging and observability
mlflow>=2.9.0           # Experiment tracking (following Spotify's approach)

# LLM Integration (example - adjust based on chosen provider)
anthropic>=0.8.0        # Claude API (Spotify uses Claude Code)
openai>=1.6.0           # OpenAI API (alternative/judge)

# PLC Integration (examples - depends on your PLCs)
# Note: These are examples, actual PLC libraries vary by vendor
# pycomm3>=1.2.0        # Allen-Bradley PLCs
# snap7>=1.3            # Siemens S7 PLCs
# pymodbus>=3.5.0       # Modbus communication

# Development and testing
pytest>=7.4.0           # Testing framework
pytest-asyncio>=0.21.0  # Async testing
black>=23.0.0           # Code formatting
mypy>=1.7.0            # Type checking

# Utilities
python-dotenv>=1.0.0    # Environment variables
requests>=2.31.0        # HTTP requests
